bugfix: resolve empty datafile error
Starting the application with an empty datafile results in an error and
causes the program to crash due to an unhandled exception. This is
because in one of my previous commits I assumed that we would receive
EOFException if the file that was being read into the objectmapper was
empty. Apparently this isn't true, and it throws a
MismatchedInputException instead. So I just added that exception to the
catch statement to resolve the issue