test: fix and add search test case
Fix the existing search test case as it didn't work before. To do this,
I rewrote it to use existing needs, and search for the string "e". Two
of the existing test needs match this criteria so it makes sure the
length of the result is correct and that these two needs specifically
are returned.

This commit also creates a new test case to test that no results are
returned upon searching for needs with a very specific string that none
of them will match