feat: implement access control and routing
Implement the routing module so that when the user navigates to the
'/cupboard' page it displays this content with the normal header and
footer.

Implpement the access control service so that this page can determine
what to show to the user based on whether they are an admin or a helper.