feat: add need
Implement the ability to add a need. This commit adds 4 input boxes
(name, type, quantity, price) above the cupboard section of the admin
page that can be used to add needs to the cupboard. The backend portion
that subscribes a reload of getting the needs again has also been
implemented. This way the needs reload when a new need is added