feat: add search
Implement search functionality so that the user can search for needs.
This required some changes to the input text box as well as recreating
the searchneeds function. This function will make a REST request
searching for the needs with the current string, unless the string is
empty in which case it will just get all of the needs. This function is
called every time there is a change in the text box (eg. character
added/removed)