feat: getpassword() function
Add getpassword() function. Originally this wasn't added for security
but I realized we will need this. This is because the credential we
receive in a REST request will need to be compared to what is on the
server and the only way to do that is be able to compare the password
field