feat: authentication controller initial commit
Create authentication controller for responding to REST requests. This
commit only implements POST requests at the mapping 'auth/login' and
'auth/'. The first POST request allows for a login request to be made,
where the request body contains the username and password. The second
POST request allows for new credentials to be stored on the server